requirements.txt
altair==4.1.0
google-api-core==2.0.1
requests==2.31.0
worldnewsapi==1.0.0
gnews==0.3.0
newsdataapi==1.0.0
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
isodate==0.6.1
Janome==0.4.1
matplotlib==3.3.1
pandas==1.3.4
pytrends==4.7.3
spotipy==2.19.0
streamlit==1.3.0
tweepy==4.4.0
youtube-dl==2021.6.6
wordcloud==1.8.1